+++ c/pq.h
@@ -20,8 +20,8 @@ int pq_less(struct pq_entry a, struct pq_entry b);
 
 struct merged_iter_pqueue {
 	struct pq_entry *heap;
-	int len;
-	int cap;
+	size_t len;
+	size_t cap;
 };
 
 struct pq_entry merged_iter_pqueue_top(struct merged_iter_pqueue pq);
